Add search bar and tags for admin order
Search bar and buttons use Ant Design. Sadly, it turns out
that the search bar might not be able to have the search icon
at the beginning, according to Ant Design. Will look into this
further.

Additionally, the tag button do not have the same colors as the
design draft for the store. Particularly, the "Completed" tag
cannot be green. For now, I've left them with sensible defaults.