Add preliminary stateless AdminOrderList
This adds a simple version of the administrator's list of orders.
Currently, only the appearance of the component is made, not its state
and actions.

This is not connected to the API as of this commit, and an example
set of test properties to pass to the list are found in
AdminOrderPage.tsx.

Similarly, the searching and filtering functionalities are not yet
implemented, however it remains to be seen how will the search be done
client-side.

This commit also raised the question of pagination to me.
Do we paginate the orders at some point? It might be prudent to do so.