Move admin order item to separate component
We need to move this to a separate file because we have to keep track
of the note modal's visibility for each item later on. Doing this inline
required a separate function scope, so as to allow for multiple visibility
variables for the Modal component.

Also, the property inside of AdminOrderItem was changed to be more clear
as to what "item" we're referring to. "item" was repeated multiple times
in the original component, "orderItem" is repeated less.