Add modal and checkbox functionality to AdminOrderItem
Modals and checkboxes now exist for each order item. They are essentially
little forms that can be used to update the relevant properties in order
items.

Currently, console statements are used to test event handlers for the modal
and checkbox. Proper actions will be used once state is implemented for the
store.