Add "orderedAt" prop to order interface in AdminOrderPage
The order date per list was not implemented with the order model in mind,
instead replaced with a static date.

Add "orderedAt" prop, as store API documentation indicates, and modify
order date for AdminOrderList to use prop. Use Date object for easy changing
of locales and date format in the future.