Pull up API functionality placeholders to AdminOrderPage container
Placeholders for actions interfacing with the API were placed in components
instead of the container, where state will reside once middleware is
implemented.

In order to ease the transition to implemented actions when Redux store is
implemented, pull up console log placeholders to AdminOrderPage container. Move
modal to AdminOrderPage to avoid duplication of the modal in AdminOrderItem.

Console log statements will be removed once actions are implemented.