UI: Admin Store Orders Table (#316)
* Add components to view all store orders in an administrator panel.
* Add middle-ground components to reach page where components are present.
* Add skeleton structure for future tagging and searching, as well as button to reach admin order page.
* Add test interfaces and properties to account for missing middleware

This page can be reached at `/admin/store/orders`, as well as following the Admin button of the sidebar. The admin order page displays each order under a specific list, with the orderer's name, profile picture, date of order, and with the individual items order and their quantities. The admin order panel can be used to fulfill items in the order, as well as adding notes and comments to each order item.