Move up local state from AdminOrderList to reducer
Most of the local state used within AdminOrderItem is rather
used due to the faulty usage of the reducer placed at the level
of the container. Move up all the local state to the reducer by
adjusting the functionality of each function and the reducer.

Additionally, move up the search function and its relevant state
from AdminOrderList to the container.

- Add action to reducer to allow for the storage of scratch notes;
  notes not yet uploaded. This avoid API hits every time one
  character is added to the TextArea of a note using Ant Design's
  component. (The component runs the callback for value change on
  every single character change, which would be devastating for
  the network bandwidth).
- Adjust reducer to allow for updating of the checkbox without
  requiring the call of FETCH_ORDERS. This occured due to the new
  items being overriden by the "...items" operator.