Add lazy-loading to Event Card images
All images on Event Cards are now lazy-loaded using external package.

This change bumps the average load time of home page from 7.29 seconds
down to 1.94 seconds, which is a massive difference. While the images
still require loading when rendered, the initial loading shock is not
as large. However, important to note that the actual image sizes should
be reduced down from 1080p, as that resolution is wasteful for the size
of the images used.

The resolution change is either mandated by the backend, or can be done
with an external script uploading old event cards to an S3 bucket at
a downscaled resolution and updating existing events using it.