Add Netlify configuration for Test Deploy Previews (#356)
We need the ability to swap API url depending
on where the UI is deployed. If deployed as a development
build, it needs to point to the testing instance.

Add 'dotenv' dependency to pass API URL depending on environment
to production build script and configure Netlify to accurately
build proper build depending on which branch it's deploying.