Add check for cheapest item price with 0% discount
In the case that an item option with 0% discount is cheaper than any
discounted option, the item card will display the cheapest price as both
a "new" and "old" struck-through price. Add check to display cheapest
option without strike-through in the case that there is no discount
applied to it.