Add Collection Item Card (#463)
* Add initial UI for item card

Note that there is no middleware for the created components.

* Add preliminary information usage in Item Card

Testing required before this works, but models on backend unclear for now.

* Adjust props and add logic for displaying prices

Adjust logic to fit specification of #457 better. Add checks for price
ranges on non-sale items and also discounts.

* Add check for cheapest item price with 0% discount

In the case that an item option with 0% discount is cheaper than any
discounted option, the item card will display the cheapest price as both
a "new" and "old" struck-through price. Add check to display cheapest
option without strike-through in the case that there is no discount
applied to it.

* Add test data for items

Fits MerchandiseItemModel as close as possible.

Fields that are extraneous and returned by the API are removed.

* Lint style files

* Rename component for CollectionItemCard

* Add link to Item Page in card container

* Pass UUID to CollectionItemCard component

* Adjust styling to add graying to whole of card

* Move types and SVG into separate files

* Fix import order for Item Card

* Adjust conditional for option price range

Account for high and low range for option prices, since this needs to
adjust in both directions if there are single option items.

* Fix linting

* Adjust number displays to include locale commas